another typo(?)

+++ source/src/sfincs_bmi.f90
@@ -501,7 +501,7 @@ contains
 
 
       character(len=strlen(var_name)) :: f_var_name
-      real(kind=c_int), pointer  :: f_inds(:)
+      integer(kind=c_int), pointer  :: f_inds(:)
       real(kind=c_float), pointer  :: f_src(:)
       integer :: i
